refactor: restructure geometry utilities
