+++ geometry/size.go
@@ -1,0 +1,6 @@
+package geometry
+
+type Size struct {
+	Width  int
+	Height int
+}
